+++ CHANGELOG.md
@@ -6,8 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- have different templates: normal/micro
-- suppprt bigger (external) bash frameworks
+- put requirements/dependencies in separate file
+
+## [1.13.8] - 2021-02-11
+### Added/changed
+- added features.md
+- MOD: README.md 
+
+## [1.13.6] - 2021-01-31
+### Added/changed
+- Update issue templates
+- Merge branch 'master' of github.com:pforret/bashew
+- ADD: CONTRIBUTING.md
 
 ## [1.1.0] - 2020-08-09
 ### Added/changed